Do not use batches anymore as batches:  1. Fails on large files as max default batchsize is 5 KB  2. Might negativly impact performance, http://tinyurl.com/jxmlcns
This change also impact file handling as now part of a file might fail.
During processing a file is renamed to ORIG_FILENAME.wip.
When finsihed processing the file is:
 1. If all is good move entire file to "processed_dir"/TODAYS-DATE/ORIG_FILENAME.json
 2. If all is bad move entire file to "failed_dir"/TODAYS-DATE/ORIG_FILENAME.json
 3. If part of file is bad:
     1. Copy the good parts to a new file in "processed_dir"/TODAYS-DATE/ORIG_FILENAME_PART1-PART2-PARTX.json, where part 1,2 etc is the good inserts
     2. Copy the bad parts to a new file in "failed_dir"/TODAYS-DATE/ORIG_FILENAME_PART3-PART4-PARTY.json, where part 3,4 etc is the bad inserts
     3. Delete original file

Action 1 and 2 are low cost while step 3 is high cost since it is a much slower process to create and write a new file than to move existing file.
Attention, if there are .wip file left in "indir" it means the importer where interuppted while doing a insert into the db.
This is a unknown state that needs manaul interaction.