Fixed a race condition on directory creation when running multiple importers
A race condition happened when multiple importers tried to create the same processed and failed directories.
In unlucky circumstances this could lead to that one of the importers threw an uncought exception and stopped.

This fix ignores file exist errors when trying to create the above directories.
Potential regression is if the directories are created outside the application without execute and write permissions.